refactor(site): inline cls variables and remove workspaceStatusDisplay
- Inline badgeCls directly into cn() call in WorkspacePill
- Remove workspaceStatusDisplay.tsx and its test file
- Inline status icon/label computation into both callers
  (WorkspacePill and AgentChatPageView)
- Replace iconCls variable with direct "size-3" on each icon
